Add logic for doing different things between ip{,6}tables
Instead of assuming, wrongly, that iptables will accept IPv6 addresses,
(which it will not) check against two regexes; if the input matches
one or the other, do the appropriate thing.  In each case, also query
the respective INPUT chain for a DROP rule on ${ip}.

So, when first checking whether something is dropped, (which is purely
a UI action at this point; all IPs are checked before being dropped)
check as described above.  When it actually comes time to drop the IP,
we now have a V6 branch on the conditional.  Now each iteration prints
what its supposed type of IP is.

Note that this does not support, for example, the syntax

    ::ffff:127.0.0.1

because this does not match either of /^[\d\.]+$/ or /^[\da-fA-F:]+$/;
namely because it includes both : and .; for this reason one should
avoid listings like this in malicious_clients.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>